refactor(checkbox): extract checkbox-group style into a separate file
